ダブルクリック

+++ app/controllers/public/comments_controller.rb
@@ -20,8 +20,15 @@ class Public::CommentsController < ApplicationController
   end
 
   def destroy
-    comment = Comment.find(params[:id])
-    comment.destroy
+    # ダブルクリック回避方法
+    park = Park.find(params[:park_id])
+    comment = params[:id]
+    park_comment = park.comments.ids
+    if park_comment.include?(comment.to_i)
+      # debugger
+       comment = Comment.find(comment)
+       comment.destroy
+    end
     redirect_to request.referer
   end
